docker/requirements.txt
numpy==1.23.5
pandas==1.5.3
psycopg2-binary==2.9.9
pyarrow==14.0.0
flask==3.0.0
